# swagger.yaml
openapi: 3.0.0
info:
  title: Nagar Parishad School Leaving Certificate API
  version: 1.0.0
  description: RESTful API for managing Nagar Parishad School Leaving Certificates with audit trail support
  contact:
    name: API Support
    email: support@nagarparishad.gov.in
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.kaamlo.com
    description: Production server
...
paths:
...
  /api/certificates/{id}/status:
    patch:
      summary: Update certificate status (Super Admin for approve/reject)
      tags: [Certificates]
      security:
        - bearerAuth: []
      description: Updates certificate status. Only Super Admin can approve/reject. Admin can submit for review.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Certificate ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [new, in_review, rejected, accepted, draft, issued, archived, cancelled]
                  example: accepted
                  description: |
                    - new: When created by Admin
                    - in_review: When submitted for review
                    - rejected: When rejected by Super Admin
                    - accepted: When approved by Super Admin
                    - draft, issued, archived, cancelled: Legacy statuses
                reason:
                  type: string
                  example: All documents verified and approved
                notes:
                  type: string
                  example: Certificate ready for issue
                user_id:
                  type: integer
                  description: User ID making the change (for audit trail)
                  example: 1
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/LeavingCertificate'
                  message:
                    type: string
                    example: Certificate status updated to accepted
        '400':
          description: Invalid status value
        '401':
          description: Not authenticated
        '403':
          description: Access denied (Only Super Admin can approve/reject)
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/certificates/{id}/status-history:
    get:
      summary: Get certificate status change history
      tags: [Certificates]
      security:
        - bearerAuth: []
      description: Returns complete history of status changes for a certificate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Certificate ID
      responses:
        '200':
          description: Status history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        certificate_id:
                          type: integer
                        old_status:
                          type: string
                        new_status:
                          type: string
                        changed_by:
                          type: integer
                        changed_by_username:
                          type: string
                        changed_by_name:
                          type: string
                        changed_at:
                          type: string
                          format: date-time
                        reason:
                          type: string
                        notes:
                          type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login or /api/auth/otp/verify
  schemas:
...
    LeavingCertificate:
      type: object
      required: [school_id, student_id, serial_no, leaving_date, leaving_class]
      properties:
        id:
          type: integer
        school_id:
          type: integer
          description: Reference to schools table
        student_id:
          type: integer
          description: Reference to students table
        serial_no:
          type: string
          example: '101'
        previous_school:
          type: string
        previous_class:
          type: string
        admission_date:
          type: string
          format: date
          example: '2019-06-01'
        admission_class:
          type: string
          example: Class 1
        progress_in_studies:
          type: string
          example: Good
        conduct:
          type: string
          example: Excellent
        leaving_date:
          type: string
          format: date
          example: '2024-03-31'
        leaving_class:
          type: string
          example: Class 5
        studying_class_and_since:
          type: string
          example: Class 5 since June 2023
        reason_for_leaving:
          type: string
          example: Transfer to another school
        remarks:
          type: string
        general_register_ref:
          type: string
          example: GR001
        certificate_date:
          type: string
          format: date
          example: '2024-04-01'
        certificate_month:
          type: string
          example: April
        certificate_year:
          type: integer
          example: 2024
        class_teacher_name:
          type: string
          example: श्रीमती सुनीता पाटिल
        clerk_name:
          type: string
          example: श्री रवींद्र जोशी
        headmaster_name:
          type: string
          example: श्री महेश देशमुख
        status:
          type: string
          enum: [new, in_review, rejected, accepted, draft, issued, archived, cancelled]
          default: new
          description: |
            - new: When created by Admin (can edit)
            - in_review: When submitted for review (can edit)
            - rejected: When rejected by Super Admin (can edit)
            - accepted: When approved by Super Admin (cannot edit, except Super Admin)
            - draft, issued, archived, cancelled: Legacy statuses
        created_by:
          type: integer
        updated_by:
          type: integer
        issued_by:
          type: integer
          description: User ID who issued the certificate
        issued_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message
        errors:
          type: array
          items:
            type: object
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Resource not found
...